# docs/api/api/openapi.yaml
openapi: 3.1.0
info:
  title: HustleX API
  description: |
    HustleX is a Nigerian fintech super app combining gig marketplace,
    social savings (Ajo/Esusu), and alternative credit scoring.

    ## Authentication

    All endpoints except `/auth/*` and `/health` require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Rate Limiting

    API requests are rate limited per endpoint category:
    - Authentication endpoints: 5 requests/minute
    - Standard endpoints: 100 requests/minute
    - Heavy endpoints (reports, exports): 10 requests/minute
    - Webhook endpoints: 1000 requests/minute

    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`: Request limit per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Pagination

    List endpoints support cursor-based pagination:
    - `limit`: Number of items per page (default: 20, max: 100)
    - `cursor`: Opaque cursor for next page (from previous response)

    Response includes:
    ```json
    {
      "data": [...],
      "pagination": {
        "next_cursor": "eyJpZCI6MTAwfQ==",
        "has_more": true
      }
    }
    ```

    ## Error Handling

    All errors follow RFC 7807 Problem Details format:
    ```json
    {
      "type": "https://api.hustlex.app/errors/insufficient-funds",
      "title": "Insufficient Funds",
      "status": 402,
...

...

...

...

...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /webhooks/paystack:
    post:
      tags: [Webhooks]
      summary: Paystack webhook
      description: |
        Webhook endpoint for Paystack events.

        Supported events:
        - charge.success
        - transfer.success
        - transfer.failed
        - transfer.reversed

        Verify webhook signature using `X-Paystack-Signature` header.
      operationId: paystackWebhook
      security: []
      parameters:
        - name: X-Paystack-Signature
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                event:
                  type: string
                data:
                  type: object
      responses:
        '200':
          description: Webhook processed
        '400':
          description: Invalid signature

  # ====================
  # HEALTH
  # ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
...